Add files via upload
Just some beginner neopixel code I'm storing on github so i dont lose it again. anyone is welcome to use and add to it. im just a beginner about 2 months in so i hope to learn alot more and actually be able to add some good stuff to the community. =)